+++ app-template-x-plus/client/src/main.x.cpp
@@ -1,18 +1,21 @@
 #include <emscripten/bind.h>
 #include <Volt.hpp>
+#include "ApplicationServices.hpp"
 #include "App.x.hpp"
 
 using namespace volt;
 using namespace emscripten;
 
-std::unique_ptr<VoltEngine> g_voltEngine = nullptr;
-
 EMSCRIPTEN_BINDINGS(VOLT_APP_NAME_UNDERSCORE_module) {
     function("getVoltNamespace", &volt::config::getVoltNamespace);
 
     function("createVoltEngine", +[](std::string rootId) {
-        g_voltEngine = std::make_unique<VoltEngine>(rootId, "VOLT_APP_NAME_UNDERSCORE_module");
-        g_voltEngine->mountApp<VOLT_APP_NAME_CAMEL>();
+        auto& application = app_detail::application();
+        if (application.engine)
+            throw std::logic_error("A Volt engine already exists in this module instance");
+        application.engine = std::make_unique<VoltEngine>(rootId, "VOLT_APP_NAME_UNDERSCORE_module");
+        services().registerDependencies(*application.engine);
+        application.engine->mountApp<VOLT_APP_NAME_CAMEL>();
     });
 
     function("invokeVoltBubbleEvent", +[](emscripten::val event) {
@@ -24,10 +27,10 @@ EMSCRIPTEN_BINDINGS(VOLT_APP_NAME_UNDERSCORE_module) {
     });
 
     function("clearVoltFocussedElements", +[]() {
-        g_voltEngine->clearFocussedElements();
+        app_detail::application().engine->clearFocussedElements();
     });
 
     function("addVoltFocussedElement", +[](emscripten::val element) {
-        g_voltEngine->addFocussedElement(element);
+        app_detail::application().engine->addFocussedElement(element);
     });
 }